qla2x00t, qla2x00t-32gbit, scst_local: Port to Linux kernel v6.14
Support for the following scsi layer changes in the Linux kernel v6.14:

- ed638918f4df ("scsi: Rename .slave_alloc() and .slave_destroy()")
- 49515b7fe50c ("scsi: Convert SCSI drivers to .sdev_configure()")
- c9a71ca13f71 ("scsi: Constify struct pci_device_id")
- bd326a5ad639 ("scsi: replace blk_mq_pci_map_queues with blk_mq_map_hw_queues")

+++ iscsi-scst/usr/chap.c
@@ -354,7 +354,7 @@ chap_calc_digest_af_alg(char *alg, char chap_id,
 	int datafd;
 	ssize_t bytes;
 	char buffer[1024];
-	int res = 0;
+	ssize_t res = 0;
 
 	datafd = af_alg_init(alg);
 	if (datafd < 0) {
